refactor: add pino and pino-pretty dependencies
This commit adds pino and pino-pretty dependencies to the package.json file. These dependencies are used for logging purposes in the application. This change enables the application to use the pino logger and format the log output using pino-pretty. This improvement enhances the logging capabilities and improves the readability of log messages.